requirements.txt
aioconsole==0.4.1
APScheduler==3.9.1
certifi==2022.5.18
distlib==0.3.4
filelock==3.7.0
keyboard==0.13.5
MetaTrader5==5.0.33
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
pipenv==2022.5.2
platformdirs==2.5.2
psutil==5.9.0
pyaes==1.6.1
pyasn1==0.4.8
pymt5adapter==0.4.4
pyparsing==3.0.9
pytesseract==0.3.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
rsa==4.8
six==1.16.0
Telethon==1.24.0
tzdata==2022.1
tzlocal==4.2
virtualenv-clone==0.5.7
wincertstore==0.2
